device: Move SPI traits
and define them on the `Inner` type. This most closely resembles an
`SpiBus`, because it has multiple CS lines attached to it, each
represented by an `Spi` struct, which adds another layer of abstraction.
Requires the `Inner` type to be public, so it can be exposed in the
`SpiDevice` trait.